+++ engine/src/main/java/net/jqwik/engine/discovery/JqwikDiscoverer.java
@@ -1,5 +1,6 @@
 package net.jqwik.engine.discovery;
 
+import java.lang.reflect.*;
 import java.util.*;
 import java.util.function.*;
 
@@ -68,10 +69,11 @@ public class JqwikDiscoverer {
 			);
 		});
 		request.getSelectorsByType(MethodSelector.class).forEach(selector -> {
+			Method javaMethod = getJavaMethodIfNecessaryFromInnerKotlinClass(selector);
 			discoveryListener.selectorProcessed(
 				engineDescriptor.getUniqueId(),
 				selector,
-				javaElementsResolver.resolveMethod(selector.getJavaClass(), selector.getJavaMethod())
+				javaElementsResolver.resolveMethod(selector.getJavaClass(), javaMethod)
 			);
 		});
 		request.getSelectorsByType(UniqueIdSelector.class).forEach(selector -> {
@@ -81,6 +83,30 @@ public class JqwikDiscoverer {
 				javaElementsResolver.resolveUniqueId(selector.getUniqueId())
 			);
 		});
+	}
+
+	private Method getJavaMethodIfNecessaryFromInnerKotlinClass(MethodSelector selector) {
+		// This is for Kotlin methods within inner classes, the names of which have a unique postfix
+		try {
+			return selector.getJavaMethod();
+		} catch (Exception methodNotFound) {
+			String methodName = selector.getMethodName();
+			for (Method method : selector.getJavaClass().getMethods()) {
+				// TODO: Also match parameters
+				if (matchesInnerKotlinMethod(method, methodName)) {
+					return method;
+				}
+			}
+			throw methodNotFound;
+		}
+	}
+
+	private boolean matchesInnerKotlinMethod(Method method, String methodName) {
+		if (!JqwikKotlinSupport.isInnerKotlinMethod(method)) {
+			return false;
+		}
+		String plainMethodName = JqwikKotlinSupport.nameWithoutSpecialPart(method);
+		return plainMethodName.equals(methodName);
 	}
 
 	private HierarchicalJavaResolver createHierarchicalResolver(TestDescriptor engineDescriptor) {